Move DeviceHelpers out of util dir.
